--- client/product/src/main/java/com/live_commerce/inventory/application/exception/CustomException.java
@@ -1,16 +1,0 @@
-package com.live_commerce.inventory.application.exception;
-
-
-import com.live_commerce.product.application.exception.ExceptionCode;
-import lombok.Getter;
-
-@Getter
-public class CustomException extends RuntimeException {
-
-  private final com.live_commerce.product.application.exception.ExceptionCode exceptionCode;
-
-  public CustomException(ExceptionCode exceptionCode) {
-    super(exceptionCode.getMessage());
-    this.exceptionCode = exceptionCode;
-  }
-}
