--- client/product/src/main/java/com/live_commerce/inventory/application/exception/ExceptionCode.java
@@ -1,10 +1,0 @@
-package com.live_commerce.inventory.application.exception;
-
-
-import org.springframework.http.HttpStatus;
-
-public interface ExceptionCode {
-
-  HttpStatus getHttpStatus();
-  String getMessage();
-}
