--- client/product/src/main/java/com/live_commerce/inventory/presentation/common/ApiResponse.java
@@ -1,13 +1,0 @@
-package com.live_commerce.inventory.presentation.common;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ApiResponse<T> {
-
-  private String status;
-  private T data;
-}
