--- client/product/src/main/java/com/live_commerce/product/application/dto/ProductCreateRequestDto.java
@@ -1,14 +1,0 @@
-package com.live_commerce.product.application.dto;
-
-import com.live_commerce.product.domain.model.ProductCategory;
-import com.live_commerce.product.domain.model.ProductStatus;
-
-import java.util.UUID;
-
-public record ProductCreateRequestDto(
-        UUID companyId,
-        String name,
-        String description,
-        Integer price,
-        ProductCategory category
-){}
