--- client/product/src/main/java/com/live_commerce/product/domain/model/BaseEntity.java
@@ -1,49 +1,0 @@
-package com.live_commerce.product.domain.model;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@Getter
-@EntityListeners(AuditingEntityListener.class)
-@MappedSuperclass
-public abstract class BaseEntity {
-
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime createdAt;
-
-    @CreatedBy
-    @Column(updatable = false)
-    private String createdBy;
-
-    @LastModifiedDate
-    private LocalDateTime updatedAt;
-
-    @LastModifiedBy
-    private String updatedBy;
-
-    private LocalDateTime deletedAt;
-
-    private String deletedBy;
-
-    private Boolean deletedStatus;
-
-    protected BaseEntity() {
-        this.deletedStatus = false;
-    }
-
-    public void delete(String deletedBy) {
-        this.deletedStatus = true;
-        this.deletedAt = LocalDateTime.now();
-        this.deletedBy = deletedBy;
-    }
-}
