--- client/product/src/main/java/com/live_commerce/product/domain/model/ProductCategory.java
@@ -1,14 +1,0 @@
-package com.live_commerce.product.domain.model;
-
-public enum ProductCategory {
-    FOOD,
-    BEAUTY,
-    FASHION,
-    DIGITAL,
-    LIVING,
-    HEALTH,
-    BABY,
-    PET,
-    BOOK,
-    ETC
-}
