--- client/product/src/main/java/com/live_commerce/product/infrastructure/client/ExternalCompanyResponseDto.java
@@ -1,13 +1,0 @@
-package com.live_commerce.product.infrastructure.client;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.UUID;
-
-
-
-public record ExternalCompanyResponseDto (
-        UUID companyId
-) { }
