--- client/product/src/main/java/com/live_commerce/product/infrastructure/config/JpaConfig.java
@@ -1,9 +1,0 @@
-package com.live_commerce.product.infrastructure.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfig {
-}
