+++ client/product/src/main/java/com/live_commerce/product/inventory/application/exception/ExceptionCode.java
@@ -1,0 +1,10 @@
+package com.live_commerce.product.inventory.application.exception;
+
+
+import org.springframework.http.HttpStatus;
+
+public interface ExceptionCode {
+
+  HttpStatus getHttpStatus();
+  String getMessage();
+}
