+++ client/product/src/main/java/com/live_commerce/product/inventory/application/exception/InventoryExceptionCode.java
@@ -1,0 +1,20 @@
+package com.live_commerce.product.inventory.application.exception;
+
+
+import com.live_commerce.product.product.application.exception.ExceptionCode;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@RequiredArgsConstructor
+public enum InventoryExceptionCode implements ExceptionCode {
+
+  NOT_FOUND(HttpStatus.NOT_FOUND, "Inventory Not Found"),
+  PRODUCT_NOT_FOUND(HttpStatus.NOT_FOUND, "상품이 존재하지 않습니다.");
+
+
+  private final HttpStatus httpStatus;
+  private final String message;
+
+}
