+++ client/product/src/main/java/com/live_commerce/product/inventory/presentation/common/ApiResponse.java
@@ -1,0 +1,13 @@
+package com.live_commerce.product.inventory.presentation.common;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ApiResponse<T> {
+
+  private String status;
+  private T data;
+}
