+++ client/product/src/main/java/com/live_commerce/product/product/application/dto/ProductUpdateRequestDto.java
@@ -1,0 +1,16 @@
+package com.live_commerce.product.product.application.dto;
+
+import com.live_commerce.product.product.domain.model.ProductCategory;
+import com.live_commerce.product.product.domain.model.ProductStatus;
+
+import java.util.UUID;
+
+public record ProductUpdateRequestDto(
+        UUID companyId,
+        String name,
+        String description,
+        Integer price,
+        ProductCategory category,
+        ProductStatus productStatus
+) {
+}
