+++ client/product/src/main/java/com/live_commerce/product/product/application/exception/CustomException.java
@@ -1,0 +1,15 @@
+package com.live_commerce.product.product.application.exception;
+
+
+import lombok.Getter;
+
+@Getter
+public class CustomException extends RuntimeException {
+
+  private final ExceptionCode exceptionCode;
+
+  public CustomException(ExceptionCode exceptionCode) {
+    super(exceptionCode.getMessage());
+    this.exceptionCode = exceptionCode;
+  }
+}
