+++ client/product/src/main/java/com/live_commerce/product/product/application/exception/ProductExceptionCode.java
@@ -1,0 +1,19 @@
+package com.live_commerce.product.product.application.exception;
+
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@RequiredArgsConstructor
+public enum ProductExceptionCode implements ExceptionCode {
+
+  NOT_FOUND(HttpStatus.NOT_FOUND, "Product Not Found"),
+  COMPANY_NOT_FOUND(HttpStatus.NOT_FOUND, "업체를 찾을 수 없습니다.");
+
+
+  private final HttpStatus httpStatus;
+  private final String message;
+
+}
