+++ client/product/src/main/java/com/live_commerce/product/product/domain/exception/ProductException.java
@@ -1,0 +1,18 @@
+package com.live_commerce.product.product.domain.exception;
+
+import com.live_commerce.product.product.application.exception.CustomException;
+import com.live_commerce.product.product.application.exception.ProductExceptionCode;
+
+public class ProductException extends CustomException {
+    public ProductException(ProductExceptionCode code) {
+        super(code);
+    }
+
+    public static void forProductNotFound() {
+        throw new ProductException(ProductExceptionCode.NOT_FOUND);
+    }
+
+    public static void forCompanyNotFound() {
+        throw new ProductException(ProductExceptionCode.COMPANY_NOT_FOUND);
+    }
+}
