+++ client/product/src/main/java/com/live_commerce/product/product/domain/model/ProductCategory.java
@@ -1,0 +1,14 @@
+package com.live_commerce.product.product.domain.model;
+
+public enum ProductCategory {
+    FOOD,
+    BEAUTY,
+    FASHION,
+    DIGITAL,
+    LIVING,
+    HEALTH,
+    BABY,
+    PET,
+    BOOK,
+    ETC
+}
