+++ client/product/src/main/java/com/live_commerce/product/product/infrastructure/client/ExternalCompanyResponseDto.java
@@ -1,0 +1,9 @@
+package com.live_commerce.product.product.infrastructure.client;
+
+import java.util.UUID;
+
+
+
+public record ExternalCompanyResponseDto (
+        UUID companyId
+) { }
