+++ client/product/src/main/java/com/live_commerce/product/product/presentation/common/ApiResponse.java
@@ -1,0 +1,13 @@
+package com.live_commerce.product.product.presentation.common;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ApiResponse<T> {
+
+  private String status;
+  private T data;
+}
